+++ packages/adminui-vue/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.6.2]
+### Fixed
+ - do not break words in the $App.dialog* messages (word-break should be break-word instead of break-all) [UBDF-9571]
+ 
 ### Changed
  - `Form->processing` changed all hooks to async, now can await result
  - `Form->processing` if beforeSave hook returns false, saving will be canceled